Move Neptune graph memory to new folder
